Сортированный связанный список. Наследование, для расширения функционала (SOLID OCP)

+++ Program/src/LinksUtils/InsertFirst.java
@@ -1,5 +1,5 @@
 package LinksUtils;
 
 public interface InsertFirst {
-    void insertFirst(int id, double dd);
+    void insertFirst(int id, int dd);
 }
